Move KubeList to parent directory
